Fix: service refactore
