+++ src/main/java/ru/yandex/practicum/filmorate/controller/MPAController.java
@@ -8,7 +8,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import ru.yandex.practicum.filmorate.model.MPA;
-import ru.yandex.practicum.filmorate.service.mpa.MPAService;
+import ru.yandex.practicum.filmorate.service.MPAService;
 import java.util.List;
 
 @RestController
