+++ src/main/java/ru/yandex/practicum/filmorate/controller/UserController.java
@@ -10,9 +10,9 @@ import ru.yandex.practicum.filmorate.exception.NotFoundException;
 import ru.yandex.practicum.filmorate.model.Event;
 import ru.yandex.practicum.filmorate.model.User;
 import ru.yandex.practicum.filmorate.exception.ValidationMarker;
-import ru.yandex.practicum.filmorate.service.event.EventService;
+import ru.yandex.practicum.filmorate.service.EventService;
 import ru.yandex.practicum.filmorate.model.Film;
-import ru.yandex.practicum.filmorate.service.user.UserService;
+import ru.yandex.practicum.filmorate.service.UserService;
 
 import java.util.List;
 import java.util.Map;
@@ -70,8 +70,7 @@ public class UserController {
 
     @GetMapping("/{id}")
     public User getUser(@PathVariable long id) {
-        return userService.getUserById(id)
-                .orElseThrow(() -> new NotFoundException("User not found"));
+        return userService.getUserById(id);
     }
 
     @GetMapping("{userId}/feed")
