+++ src/main/java/ru/yandex/practicum/filmorate/service/DirectorService.java
@@ -1,4 +1,4 @@
-package ru.yandex.practicum.filmorate.service.director;
+package ru.yandex.practicum.filmorate.service;
 
 
 import lombok.RequiredArgsConstructor;
