+++ src/main/java/ru/yandex/practicum/filmorate/service/EventService.java
@@ -1,4 +1,4 @@
-package ru.yandex.practicum.filmorate.service.event;
+package ru.yandex.practicum.filmorate.service;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
